Convert site/menu.js to typescript
