+++ index.html
@@ -386,7 +386,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 	</div>
 </div>
 
-<script type="text/javascript" src="site/menu.js"></script>
 <script type="text/javascript" src="dist/demo.js"></script>
 </body>
 </html>
